+++ backend/src/validators/report.js
@@ -2,7 +2,7 @@ const { Schema, getResponse } = require("./common_schema.js");
 const S = require("fluent-json-schema");
 
 const getOne = {
-  params: S.object().prop("projectId", Schema.Id),
+  params: Schema.IdParam,
   response: getResponse(S.object().prop("projectId", Schema.Id)),
 };